+++ tests/mypy/outputs/plugin-success-strict.txt
@@ -1,3 +1,3 @@
-28: error: Unexpected keyword argument "z" for "Model"  [call-arg]
-63: error: Untyped fields disallowed  [pydantic-field]
-78: error: Argument "x" to "OverrideModel" has incompatible type "float"; expected "int"  [arg-type]
+29: error: Unexpected keyword argument "z" for "Model"  [call-arg]
+64: error: Untyped fields disallowed  [pydantic-field]
+79: error: Argument "x" to "OverrideModel" has incompatible type "float"; expected "int"  [arg-type]
