+++ tests/mypy/outputs/success.txt
@@ -1,0 +1,3 @@
+119: error: Unexpected keyword argument "name" for "AddProject"  [call-arg]
+119: error: Unexpected keyword argument "slug" for "AddProject"  [call-arg]
+119: error: Unexpected keyword argument "description" for "AddProject"  [call-arg]
