+++ tests/mypy/test_mypy.py
@@ -21,12 +21,11 @@ cases = [
     ('mypy-plugin.ini', 'plugin_fail.py', 'plugin-fail.txt'),
     ('mypy-plugin-strict.ini', 'plugin_success.py', 'plugin-success-strict.txt'),
     ('mypy-plugin-strict.ini', 'plugin_fail.py', 'plugin-fail-strict.txt'),
-    ('mypy-default.ini', 'success.py', None),
+    ('mypy-default.ini', 'success.py', 'success.txt'),
     ('mypy-default.ini', 'fail1.py', 'fail1.txt'),
     ('mypy-default.ini', 'fail2.py', 'fail2.txt'),
     ('mypy-default.ini', 'fail3.py', 'fail3.txt'),
-    ('mypy-default.ini', 'fail4.py', 'fail4.txt'),
-    ('mypy-default.ini', 'plugin_success.py', None),
+    ('mypy-default.ini', 'plugin_success.py', 'plugin-success.txt'),
 ]
 executable_modules = list({fname[:-3] for _, fname, out_fname in cases if out_fname is None})
